+++ waterfun-service-core/src/main/java/org/waterwood/waterfunservicecore/infrastructure/persistence/TagRepository.java
@@ -1,11 +1,9 @@
 package org.waterwood.waterfunservicecore.infrastructure.persistence;
 
-import org.springframework.data.domain.Page;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
-import org.waterwood.api.VO.OptionVO;
 import org.waterwood.waterfunservicecore.entity.post.Tag;
 import org.waterwood.common.jpa.SlugUniquenessChecker;
 
@@ -13,14 +11,14 @@ import java.util.Collection;
 import java.util.List;
 import java.util.Set;
 
-public interface TagRepository extends JpaRepository<Tag, Integer>, JpaSpecificationExecutor<Tag>, SlugUniquenessChecker {
+public interface TagRepository extends JpaRepository<Tag, Long>, JpaSpecificationExecutor<Tag>, SlugUniquenessChecker {
   boolean existsTagBySlug(String slug);
 
   List<Tag> findAllByCreatorUid(Long currentUserUid);
 
-  List<Integer> findTagIdsByTagsIdIn(Collection<Integer> tagsIds);
+  List<Long> findTagIdsByTagsIdIn(Collection<Long> tagsIds);
 
-  int removeByIdIn(Collection<Integer> ids);
+  int removeByIdIn(Collection<Long> ids);
 
   @Query("SELECT t FROM Tag t ORDER BY t.usageCount DESC LIMIT :limit")
   List<Tag> findTopByOrderByUsageCountDesc(@Param("limit") Integer limit);
